refactor(workflow-tasks): move AnsibleAdapter into library, drop ToolPaths coupling
