+++ tools/workflow-tasks/tests/infra/test_ansible.py
@@ -1,0 +1,33 @@
+from __future__ import annotations
+
+from pathlib import Path
+
+from workflow_tasks.infra.ansible import AnsibleAdapter
+from workflow_tasks.shell import RecordingShell
+from workflow_tasks.vm.models import VmRequest
+
+
+def test_provision_base_uses_ops_ansible_root() -> None:
+    shell = RecordingShell()
+    adapter = AnsibleAdapter(repo_root=Path("/repo"), shell=shell)
+    request = VmRequest(lifecycle="external", host="vm.example.test", user="dev")
+
+    adapter.provision_base(request, dry_run=True)
+
+    command = shell.commands[0]
+    assert "ansible-playbook" in command
+    assert "ops/ansible/playbooks/provision-base.yml" in " ".join(command)
+    assert "vm.example.test," in command
+
+
+def test_configure_k3s_registry_sets_expected_extra_vars() -> None:
+    shell = RecordingShell()
+    adapter = AnsibleAdapter(repo_root=Path("/repo"), shell=shell)
+    request = VmRequest(lifecycle="external", host="vm.example.test", user="dev")
+
+    adapter.configure_k3s_registry(request, registry="registry.example.test:5000", dry_run=True)
+
+    rendered = " ".join(shell.commands[0])
+    assert "configure-k3s-registry.yml" in rendered
+    assert "registry=registry.example.test:5000" in rendered
+    assert "registry_port=5000" in rendered
